RUN 38: +321 insn, 55.98% -> 56.15% (106,831 / 190,254)
mini_fight lbl_00006A40 (321) -- a never-drafted, never-named blank page.
0 in 0 on words AND relocations, TRUE positional 321/321, refs 70/70,
MISMATCHED 0, no FICTION rows; objdump -h gives .text 0x0504 and nothing
else, so zero layout cost is proved by section report.

Eight of nine modules banked zero. Five landed a declarator precondition
that banks nothing and gates GOLDEN:

  src/mini_bowling_4bb.c   int -> s8 return type on a MATCHED function,
                           both occurrences (a fifth declarator class)
  src/mini_golf_51.c/_68.c wrong return type AND wrong params, not K&R
  src/mini_race_28m.c      forward decl and asm stub declarator
  src/sel_ngc_rel_29.c     extern void -> int is_minigame_unlocked(int)
  src/test_mode_112.c      wrong SIGNATURE vs a header typedef; the
                           declaration and the asm void definition both

tools/rel_relscore.py db56c6d7 -> fa37bff0 (LF preserved)
tools/rel_tuprobe.py  d9a7c08c -> c629ae5f (CRLF preserved)

Wires the peephole regime into rel_relscore -- 23 of 162 live rows were
being scored in the wrong regime -- and implements the multi-body abort in
BOTH body loops, which run 37 specified and deliberately declined.

Three gates, re-run against the POST-merge tree, landed by direct copy
because core.autocrlf=true makes git apply itself flip LF->CRLF:
  gate_regime     ARM A 162/162, ARM B 23/23 FIRED, ARM C 139/139
  gate_abort      both body loops, exit code stays 2
  gate_directive  10 cases as recorded; SILENT PARTIAL still confirmed
                  on //@SUB and //@PROTO second occurrences (not fixed)

Verified in the main tree, not taken on report: all nine gate GOLDEN from
deleted objects with every hash equal to its own agent's; all nine
rel_structcheck CLEAN; clean build from 0 objects and 0 artifacts gives
make exit status 0 and sha1sum -c 12/12 OK including the DOL.
Object metrics 784 / 954 / 789 unchanged -- no carve, no TU merge.

+++ src/mini_bowling_4bb.c
@@ -296,7 +296,7 @@ struct BowlConfig
     u8 pad0[0x1fcc];
     struct PatEntry arr2[27];
 };
-int lbl_0000664C(u8 *p, s8 i8);
+s8 lbl_0000664C(u8 *p, s8 i8);
 void u_play_sound_0(int arg0);
 void u_play_music(u32 arg0, s8 arg1);
 
@@ -1723,7 +1723,7 @@ asm void lbl_00005B0C(void)
 }
 
 #pragma peephole on
-int lbl_0000664C(u8 *p, s8 i8)
+s8 lbl_0000664C(u8 *p, s8 i8)
 {
     u8 *q;
     int i;
